Rename WaitForPartitionSplittingQueue
